hw2 2024 update

+++ python/needle/init/init_basic.py
@@ -16,11 +16,14 @@ def randn(*shape, mean=0.0, std=1.0, device=None, dtype="float32", requires_grad
     return ndl.Tensor(array, device=device, dtype=dtype, requires_grad=requires_grad)
 
 
+
+
 def constant(*shape, c=1.0, device=None, dtype="float32", requires_grad=False):
     """Generate constant Tensor"""
     device = ndl.cpu() if device is None else device
     array = device.ones(*shape, dtype=dtype) * c  # note: can change dtype
     return ndl.Tensor(array, device=device, dtype=dtype, requires_grad=requires_grad)
+
 
 
 def ones(*shape, device=None, dtype="float32", requires_grad=False):
@@ -48,7 +51,7 @@ def one_hot(n, i, device=None, dtype="float32", requires_grad=False):
     """Generate one-hot encoding Tensor"""
     device = ndl.cpu() if device is None else device
     return ndl.Tensor(
-        device.one_hot(n, i.numpy(), dtype=dtype),
+        device.one_hot(n, i.numpy().astype("int32"), dtype=dtype),
         device=device,
         requires_grad=requires_grad,
     )
